// client/app/api/accounting/model/AccountBookModel.ts
/**
 * GuanPlus.AccountingFirm.WebApi
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import * as models from './models';

/**
 * 账套
 */
export interface AccountBookModel {
    /**
     * 公司id
     */
    id?: string;

    /**
     * 账套状态
     */
    status?: AccountBookModel.StatusEnum;

    /**
     * 账套名字
     */
    name?: string;

    /**
     * 法人
     */
    legalPersonName?: string;

    /**
     * 注册资金
     */
    registeredCapital?: number;

    /**
     * 成立日期
     */
    beginDate?: Date;

    /**
     * 税号
     */
    taxNumber?: string;

    /**
     * 公司性质
     */
    companyProperty?: AccountBookModel.CompanyPropertyEnum;

    /**
     * 公司所属行业
     */
    industry?: models.IdNameModel;

    /**
     * 开始记账月份
     */
    openingDate?: Date;

    /**
     * 当前记账月份
     */
    currentDate?: Date;

    /**
     * 服务截至日期
     */
    serviceDeadline?: Date;

    /**
     * 开票金额
     */
    issueInvoiceAmount?: number;

    /**
     * 收票金额
     */
    receiveInvoiceAmount?: number;

    /**
     * 会计
     */
    account?: models.AccountBookUserModel;

    /**
     * 助理
     */
    assistant?: models.AccountBookUserModel;

    /**
     * 客户联系人
     */
    customer?: models.AccountBookUserModel;

    /**
     * 地址
     */
    address?: models.AddressModel;

    /**
     * 备注
     */
    remark?: string;

    /**
     * 期初导航
     */
    beginningNavigation?: models.AccountBookBeginningNavigationEnumModel;

    /**
     * 导入excel中对应的行数
     */
    importRow?: number;

    /**
     * ImportTaskId
     */
    importTaskId?: string;

}
export namespace AccountBookModel {
    export enum StatusEnum {
        BeginningInit = <any> 'BeginningInit',
        InProgress = <any> 'InProgress',
        CarryForward = <any> 'CarryForward',
        Posting = <any> 'Posting',
        Done = <any> 'Done'
    }
    export enum CompanyPropertyEnum {
        SmallScale = <any> 'SmallScale',
        GeneralTaxpayer = <any> 'GeneralTaxpayer'
    }
}