// client/app/api/identity/api/InvoiceApi.ts
export class InvoiceApi {
    protected basePath = 'http://guanplus-api-identity-dev.cn-north-1.eb.amazonaws.com.cn';
    public defaultHeaders: Headers = new Headers();

    constructor(protected http: Http, @Optional()@Inject(IDENTITY_BASE_PATH) basePath: string) {
        if (basePath) {
            this.basePath = basePath;
        }
    }

    /**
     * 生成发票
     * 
     * @param invoiceModel 
     */
    public invoiceCreate (invoiceModel: models.InvoiceModel,
      extraHttpRequestParams?: any ): Observable<models.InvoiceModel> {
        const path = this.basePath + '/api/v1/invoice';

        let queryParameters = new URLSearchParams();
        let headerParams = this.defaultHeaders;
        // verify required parameter 'invoiceModel' is not null or undefined
        if (invoiceModel === null || invoiceModel === undefined) {
            throw new Error('Required parameter invoiceModel was null or undefined' +
              'when calling invoiceCreate.');
        }
        let requestOptions: RequestOptionsArgs = {
            method: RequestMethod.Post,
            headers: headerParams,
            search: queryParameters
        };
        requestOptions.body = JSON.stringify(invoiceModel);

        return this.http.request(path, requestOptions)
            .map((response: Response) => {
                if (response.status === 204) {
                    return undefined;
                } else {
                    return response.json();
                }
            });
    }

    /**
     * 获取发票列表
     * 
     * @param companyId 
     * @param pageIndex 
     * @param pageSize 
     */
    public invoiceSearch (companyId: string,
      pageIndex?: string,
      pageSize?: string,
      extraHttpRequestParams?: any ): Observable<models.PagedResultInvoiceModel> {
        const path = this.basePath + '/api/v1/invoice/search';

        let queryParameters = new URLSearchParams();
        let headerParams = this.defaultHeaders;
        // verify required parameter 'companyId' is not null or undefined
        if (companyId === null || companyId === undefined) {
            throw new Error('Required parameter companyId was null or undefined' +
              'when calling invoiceSearch.');
        }
        if (companyId !== undefined) {
            queryParameters.set('companyId', String(companyId));
        }

        if (pageIndex !== undefined) {
            queryParameters.set('pageIndex', String(pageIndex));
        }

        if (pageSize !== undefined) {
            queryParameters.set('pageSize', String(pageSize));
        }

        let requestOptions: RequestOptionsArgs = {
            method: RequestMethod.Get,
            headers: headerParams,
            search: queryParameters
        };

        return this.http.request(path, requestOptions)
            .map((response: Response) => {
                if (response.status === 204) {
                    return undefined;
                } else {
                    return response.json();
                }
            });
    }

}